+++ src/main/scala/co/topl/brambl/cli/BramblCliParamsParserModule.scala
@@ -41,11 +41,16 @@ object BramblCliParamsParserModule {
 
   val amountArg = opt[Long]('a', "amount")
     .action((x, c) => c.copy(amount = x))
-    .text("Amount to send or mint")
+    .text("Amount to send")
     .validate(x =>
       if (x > 0) success
       else failure("Amount must be greater than 0")
     )
+
+  val mintAmountArg = opt[Long]("mint-amount")
+    .action((x, c) => c.copy(amount = x))
+    .text("Amount to mint")
+    .optional()
 
   val newwalletdbArg = opt[String]("newwalletdb")
     .action((x, c) => c.copy(walletFile = x))
@@ -464,7 +469,7 @@ object BramblCliParamsParserModule {
               )
           )) ++
             Seq(
-              amountArg,
+              mintAmountArg,
               opt[Long]("fee")
                 .action((x, c) => c.copy(fee = x))
                 .text("Fee paid for the transaction")
@@ -482,10 +487,33 @@ object BramblCliParamsParserModule {
                   c.tokenType != TokenType.asset
                 )
                   failure(
-                    "Only group, series and asset minting is supported at the moment"
+                    "Invalid asset to mint, supported assets are group, series and asset"
                   )
-                else
-                  success
+                else {
+                  if (
+                    c.mode == BramblCliMode.simpleminting &&
+                    c.subcmd == BramblCliSubCmd.create
+                  ) {
+                    if (c.tokenType == TokenType.asset) {
+                      if (c.amount < 0) { // not set
+                        success
+                      } else {
+                        failure(
+                          "Amount is only mandatory for group and series minting"
+                        )
+                      }
+                    } else {
+                      if (c.amount > 0) {
+                        success
+                      } else {
+                        failure(
+                          "Amount is mandatory for group and series minting"
+                        )
+                      }
+                    }
+                  } else
+                    success
+                }
               )
             )): _*
         )